requirements.txt
pandas
python-dotenv
numpy
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
pyaml
scipy

# -e .